+++ ot3_state_manager/pyproject.toml
@@ -5,12 +5,11 @@ description = "State Manager for OT3"
 license = "Apache-2.0"
 repository = "https://github.com/Opentrons/ot3-firmware"
 authors = ["Opentrons <engineering@opentrons.com>"]
-packages = [
-    { include = "header_generation_utils", from = "../"}
-]
 
 [tool.poetry.dependencies]
 python = "^3.8"
+
+
 
 [tool.poetry.dev-dependencies]
 pytest = "7.0.1"
@@ -31,6 +30,7 @@ mdformat_footnote = "^0.1.1"
 # Expects the monorepo (opentrons) to be a sibling of this repo
 opentrons-hardware = { path = "../../opentrons/hardware" }
 opentrons = { path = "../../opentrons/api" }
+header_generation_utils = { path = "../header_generation_utils"}
 
 [tool.pytest.ini_options]
 addopts = "--color=yes"
